Fetch all submitters at the same time
Instead of making a lot of request for the individual submitters, make a
single request for them all. This means a larger initial load and
possibly fetching submitters which aren't referred to by flow binders,
but this is preferable to the performance degradation which happens when
javascript client side makes several requests  (especially since there is
no multi-threading in javascript).